+++ src/main/java/com/pepper/project/he/article/mapper/ArticleMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ArticleMapper {
 
-    Article selectArticleById(Integer id);
+    Article selectArticleById(Long id);
     List<Article> selectArticleList(Article article);
 
     public int insertArticle(Article board);
