+++ src/main/java/com/pepper/project/he/video/service/IVideoService.java
@@ -11,7 +11,7 @@ public interface IVideoService {
 
     int deleteVideoByIds(String ids);
 
-    Video selectVideoById(Integer id);
+    Video selectVideoById(Long id);
 
     int updateVideo(Video video);
